Move serialization and traversals stuff
